+++ plot/viewer/src/canvases/SketchCanvas.tsx
@@ -366,7 +366,7 @@ function SketchCanvasInner({
 
   // ---------------- Context menu ----------------
 
-  const { menu, closeMenu, openNodeMenu, openEdgeMenu, openPaneMenu } = useContextMenus({
+  const { menu, closeMenu, openNodeMenu, openEdgeMenu, openPaneMenu, onNodeDragStop } = useContextMenus({
     docRef,
     flowRef,
     onDocChange,
@@ -446,6 +446,7 @@ function SketchCanvasInner({
         onNodeContextMenu={openNodeMenu}
         onEdgeContextMenu={openEdgeMenu}
         onEdgeDoubleClick={(_evt, edge) => setEdgeModalId(edge.id)}
+        onNodeDragStop={onNodeDragStop}
         onNodeClick={handleNodeClick}
         onNodeDoubleClick={handleNodeDoubleClick}
         onPaneClick={handlePaneClick}